feat: rename root into core
