+++ src/libs/satellite/src/sdk/core/hooks.rs
