+++ src/libs/satellite/src/sdk/mod.rs
@@ -1,2 +1,2 @@
+pub mod core;
 pub mod internal;
-pub mod root;
